requirements.txt
# Core functionality
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
sentence-transformers==4.0.1
slack-sdk==3.35.0
python-dotenv==1.1.0
pytz==2024.2
python-dateutil==2.9.0.post0 
Jinja2==3.1.4

# Web framework
Flask==3.0.3

# Asynchronous task processing
celery==5.5.0
redis==5.2.1